refactor: bring back ViewModeSwitcher
